Move the trio gunicorn worker to it's own file.
